refactor(header): V1 phase 5bb — split 322-line Header into focused parts
Master-piece review #11 flagged Header.tsx as one of the worst
prop-drilling / monolith examples in the frontend. One file, 322
lines, mixing navigation, user menu, mobile drawer, brand link,
nav-link primitives, and the active-underline animation layer.

Replaced with a composer at the top + five focused subcomponents in
a new `layout/header/` directory:

* `Header.tsx` (77 lines) — owns one piece of state (`mobileOpen`)
  and the cross-cutting 'close on route change' effect. Composes the
  brand link inline (two lines, no internal state worth extracting)
  with four child components.
* `header/HeaderNavLink.tsx` (75 lines) — shared link primitive
  used by both desktop and sheet variants; exports the LayoutGroup
  re-export and the shared underline layout id so consumers don't
  reach into motion/react directly.
* `header/HeaderDesktopNav.tsx` (60 lines) — `>= md` top-bar nav.
  Pulls the active-route logic locally from useLocation; the parent
  no longer threads `isActive` through three layers of props.
* `header/HeaderUserMenu.tsx` (90 lines) — `>= md` right cluster.
  Bell (lazy-loaded) + avatar / profile for authed users, sign-in +
  register for guests.
* `header/HeaderMobileMenuTrigger.tsx` (45 lines) — `< md` hamburger
  button. State lifted to the parent for the close-on-route-change
  effect.
* `header/HeaderMobileSheet.tsx` (123 lines) — the entire drawer.
  Mirrors the desktop nav with the verbose label variants per the
  UI-DECISIONS doc.

Total LOC: 322 → 77 (composer) + 393 (subcomponents) = 470 spread
across 6 focused files instead of one monolith. Per-file
responsibilities are now single-paragraph describable.

Doc-comments on each subcomponent explain WHY the seam is where it
is — the compact-vs-verbose label split between desktop and sheet,
the inline panel of the bell on mobile (Phase 5ah), the lazy bell
load to keep the unauthed home-page bundle small.

No behavioural change. App.tsx import path unchanged (`Header`
still re-exports as the layout/Header default). 323 frontend tests
pass. lint + tsc clean.
